Add note command parser test cases
